+++ mex/extractors/confluence_vvt/main.py
@@ -1,6 +1,8 @@
 from itertools import tee
 from typing import Any
 
+from dagster import asset
+
 from mex.common.cli import entrypoint
 from mex.common.ldap.extract import get_merged_ids_by_query_string
 from mex.common.ldap.transform import transform_ldap_persons_with_query_to_mex_persons
@@ -24,7 +26,7 @@ from mex.extractors.confluence_vvt.transform import (
     transform_confluence_vvt_activities_to_extracted_activities,
 )
 from mex.extractors.filters import filter_by_global_rules
-from mex.extractors.pipeline import asset, run_job_in_process
+from mex.extractors.pipeline import run_job_in_process
 from mex.extractors.settings import Settings
 from mex.extractors.sinks import load
 from mex.extractors.utils import load_yaml