+++ mex/extractors/endnote/main.py
@@ -1,3 +1,5 @@
+from dagster import asset
+
 from mex.common.cli import entrypoint
 from mex.common.models import (
     BibliographicResourceMapping,