+++ mex/extractors/grippeweb/main.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 from typing import Any
 
+from dagster import asset
+
 from mex.common.cli import entrypoint
 from mex.common.ldap.transform import (
     transform_ldap_actors_to_mex_contact_points,
@@ -37,7 +39,7 @@ from mex.extractors.grippeweb.transform import (
     transform_grippeweb_variable_group_to_extracted_variable_groups,
     transform_grippeweb_variable_to_extracted_variables,
 )
-from mex.extractors.pipeline import asset, run_job_in_process
+from mex.extractors.pipeline import run_job_in_process
 from mex.extractors.settings import Settings
 from mex.extractors.sinks import load
 from mex.extractors.sumo.transform import get_contact_merged_ids_by_emails
@@ -166,7 +168,7 @@ def extracted_access_platform_grippeweb(
 
 
 @asset(group_name="grippeweb")
-def grippeweb_extracted_resource_dict(  # noqa: PLR0913
+def grippeweb_extracted_parent_resource(  # noqa: PLR0913
     grippeweb_resource_mappings: list[dict[str, Any]],
     unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
     extracted_access_platform_grippeweb: ExtractedAccessPlatform,
@@ -174,26 +176,29 @@ def grippeweb_extracted_resource_dict(  # noqa: PLR0913
     extracted_mex_persons_grippeweb: list[ExtractedPerson],
     grippeweb_organization_ids_by_query_string: dict[str, MergedOrganizationIdentifier],
     extracted_mex_functional_units_grippeweb: dict[Email, MergedContactPointIdentifier],
-) -> dict[str, ExtractedResource]:
+) -> ExtractedResource:
     """Transform Grippeweb default values to extracted resources and load to sinks."""
-    extracted_resources = transform_grippeweb_resource_mappings_to_extracted_resources(
-        [ResourceMapping.model_validate(r) for r in grippeweb_resource_mappings],
-        unit_stable_target_ids_by_synonym,
-        extracted_access_platform_grippeweb,
-        extracted_primary_source_grippeweb,
-        extracted_mex_persons_grippeweb,
-        grippeweb_organization_ids_by_query_string,
-        extracted_mex_functional_units_grippeweb,
+    parent_resource, child_resource = (
+        transform_grippeweb_resource_mappings_to_extracted_resources(
+            [ResourceMapping.model_validate(r) for r in grippeweb_resource_mappings],
+            unit_stable_target_ids_by_synonym,
+            extracted_access_platform_grippeweb,
+            extracted_primary_source_grippeweb,
+            extracted_mex_persons_grippeweb,
+            grippeweb_organization_ids_by_query_string,
+            extracted_mex_functional_units_grippeweb,
+        )
     )
-    load(list(extracted_resources.values()))
-    return extracted_resources
+    load([parent_resource])
+    load([child_resource])
+    return parent_resource
 
 
 @asset(group_name="grippeweb")
 def grippeweb_extracted_variable_group(
     grippeweb_variable_group: dict[str, Any],
     grippeweb_columns: dict[str, dict[str, list[Any]]],
-    grippeweb_extracted_resource_dict: dict[str, ExtractedResource],
+    grippeweb_extracted_parent_resource: ExtractedResource,
     extracted_primary_source_grippeweb: ExtractedPrimarySource,
 ) -> list[ExtractedVariableGroup]:
     """Transform Grippeweb values to extracted variable groups and load to sinks."""
@@ -201,7 +206,7 @@ def grippeweb_extracted_variable_group(
         transform_grippeweb_variable_group_to_extracted_variable_groups(
             VariableGroupMapping.model_validate(grippeweb_variable_group),
             grippeweb_columns,
-            grippeweb_extracted_resource_dict,
+            grippeweb_extracted_parent_resource,
             extracted_primary_source_grippeweb,
         )
     )
@@ -214,7 +219,7 @@ def grippeweb_extracted_variable(
     grippeweb_variable: dict[str, Any],
     grippeweb_extracted_variable_group: list[ExtractedVariableGroup],
     grippeweb_columns: dict[str, dict[str, list[Any]]],
-    grippeweb_extracted_resource_dict: dict[str, ExtractedResource],
+    grippeweb_extracted_parent_resource: ExtractedResource,
     extracted_primary_source_grippeweb: ExtractedPrimarySource,
 ) -> None:
     """Transform Grippeweb default values to extracted variables and load to sinks."""
@@ -222,7 +227,7 @@ def grippeweb_extracted_variable(
         VariableMapping.model_validate(grippeweb_variable),
         grippeweb_extracted_variable_group,
         grippeweb_columns,
-        grippeweb_extracted_resource_dict,
+        grippeweb_extracted_parent_resource,
         extracted_primary_source_grippeweb,
     )
     load(extracted_variables)