+++ mex/extractors/international_projects/extract.py
@@ -194,7 +194,7 @@ def extract_international_projects_partner_organizations(
 
 
 def get_temporal_entity_from_cell(
-    cell_value: Any,
+    cell_value: Any,  # noqa: ANN401
 ) -> TemporalEntity | YearMonthDay | None:
     """Try to extract a temporal_entity from a cell.