+++ tests/biospecimen/test_main.py
@@ -5,5 +5,4 @@ from mex.extractors.pipeline.base import run_job_in_process
 
 @pytest.mark.usefixtures("mocked_ldap", "mocked_wikidata")
 def test_job() -> None:
-    result = run_job_in_process("biospecimen")
-    assert result.success
+    assert run_job_in_process("biospecimen")